Move KVError from pkg/kp to pkg/kp/consulutil
The KVError type wraps around Consul's errors, hiding the raw error message and
adding debug info. This is Consul-specific behavior, and this change moves it
to a Consul-specific utility library.
